M5 hw1 fixes after review
